:recycle: refactor: Reorder function calls at post service entrypoint

+++ post/service/service_impl.go
@@ -9,16 +9,16 @@ import (
 )
 
 var (
-	errOutOfRangeLimit  = errors.New("limit is out of range")
-	errOutOfRangeOffset = errors.New("offset is out of range")
-	errOutOfRangeId     = errors.New("id is out of range")
+	ErrOutOfRangeLimit  = errors.New("`limit` must be a positive integer")
+	ErrOutOfRangeOffset = errors.New("`offset` must be an integer bigger than -1")
+	ErrOutOfRangeId     = errors.New("`id` must be a positive integer")
 )
 
 type postService struct {
 	repository repository.PostRepository
 }
 
-func NewPostService(repository repository.PostRepository) PostService {
+func NewPostService(repository repository.PostRepository) *postService {
 	return &postService{
 		repository: repository,
 	}
@@ -26,11 +26,11 @@ func NewPostService(repository repository.PostRepository) PostService {
 
 func (svc *postService) GetPosts(ctx context.Context, pagination *entity.Pagination) ([]*entity.Post, error) {
 	if pagination.Limit < 1 {
-		return nil, errOutOfRangeLimit
+		return nil, ErrOutOfRangeLimit
 	}
 
 	if pagination.Offset < 0 {
-		return nil, errOutOfRangeOffset
+		return nil, ErrOutOfRangeOffset
 	}
 
 	posts, err := svc.repository.GetPosts(ctx, pagination)
@@ -44,7 +44,7 @@ func (svc *postService) GetPosts(ctx context.Context, pagination *entity.Paginat
 
 func (svc *postService) GetPostById(ctx context.Context, id int) (*entity.Post, error) {
 	if id < 1 {
-		return nil, errOutOfRangeId
+		return nil, ErrOutOfRangeId
 	}
 
 	post, err := svc.repository.GetPostById(ctx, id)
@@ -67,6 +67,10 @@ func (svc *postService) CreatePost(ctx context.Context, data *entity.Post) (*ent
 }
 
 func (svc *postService) DeletePost(ctx context.Context, postId int) (*entity.Post, error) {
+	if postId < 0 {
+		return nil, ErrOutOfRangeId
+	}
+
 	post, err := svc.repository.DeletePost(ctx, postId)
 
 	if err != nil {
